+++ packages/blocks/src/api/index.js
@@ -20,7 +20,7 @@ export {
 	getSaveElement,
 	getSaveContent,
 } from './serializer';
-export { isBlockContentValid, isValidBlock } from './validation';
+export { isValidBlockContent, isValidBlock } from './validation';
 export {
 	getCategories,
 	setCategories,