+++ packages/blocks/src/api/test/validation.js
@@ -13,7 +13,7 @@ import {
 	isEqualTokensOfType,
 	getNextNonWhitespaceToken,
 	isEquivalentHTML,
-	isBlockContentValid,
+	isValidBlockContent,
 	isClosedByToken,
 } from '../validation';
 import {
@@ -553,11 +553,11 @@ describe( 'validation', () => {
 		} );
 	} );
 
-	describe( 'isBlockContentValid()', () => {
+	describe( 'isValidBlockContent()', () => {
 		it( 'returns false if block is not valid', () => {
 			registerBlockType( 'core/test-block', defaultBlockSettings );
 
-			const isValid = isBlockContentValid(
+			const isValid = isValidBlockContent(
 				getBlockType( 'core/test-block' ),
 				{ fruit: 'Bananas' },
 				'Apples'
@@ -576,7 +576,7 @@ describe( 'validation', () => {
 				},
 			} );
 
-			const isValid = isBlockContentValid(
+			const isValid = isValidBlockContent(
 				getBlockType( 'core/test-block' ),
 				{ fruit: 'Bananas' },
 				'Bananas'
@@ -589,7 +589,7 @@ describe( 'validation', () => {
 		it( 'returns true is block is valid', () => {
 			registerBlockType( 'core/test-block', defaultBlockSettings );
 
-			const isValid = isBlockContentValid(
+			const isValid = isValidBlockContent(
 				getBlockType( 'core/test-block' ),
 				{ fruit: 'Bananas' },
 				'Bananas'